Rename user_invite_service file
