+++ backend/python/app/dependencies/services.py
@@ -34,7 +34,7 @@ from app.services.implementations.route_group_service import RouteGroupService
 from app.services.implementations.scheduler_service import SchedulerService
 from app.services.implementations.system_settings_service import SystemSettingsService
 from app.services.implementations.user_service import UserService
-from app.services.implementations.user_invite_services import UserInviteService
+from app.services.implementations.user_invite_service import UserInviteService
 from app.services.implementations.user_service import UserService
 from app.services.protocols.routing_algorithm import RoutingAlgorithmProtocol
 from app.utilities.gcp_client import GCPStorageClient
